+++ lib/flutter_repository_sqflite.dart
@@ -5,6 +5,7 @@ import 'package:sqflite/sqflite.dart';
 import 'src/data_source.dart';
 import 'src/migration.dart';
 
+export 'src/persistence.dart';
 export 'src/exceptions.dart';
 export 'src/migration.dart';
 export 'src/version.dart';