+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [1.1.1]
+Make previously introduced classes publicly available.
+
+## [1.1.0]
+Add Persistence and ApplicationPersistence classes to organize
+persistence-related code in mid-to-large applications.
+
 ## [1.0.2]
 Add equality check support to Version.